requirements.txt
streamlit
requests
beautifulsoup4
fpdf2
python-whois